Refine map path normalization and add regression test
